refactor: delete redux-toolkit + axios code
