+++ src/app/providers/store/config/hooks.ts
@@ -1,7 +1,7 @@
 import { TypedUseSelectorHook, useDispatch } from "react-redux";
 import { useSelector } from "react-redux";
 
-import { AppDispatch, RootState } from "../store";
+import { AppDispatch, RootState } from "./store";
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
