+++ src/app/providers/store/config/store.ts
@@ -1,12 +1,7 @@
 import { combineReducers, configureStore } from "@reduxjs/toolkit";
 import { characterApi } from "src/shared/api/charactersApi";
 
-import { charactersReducer } from "./reducers/CharactersSlice";
-import { oneCharacterReducer } from "./reducers/OneCharacterSlice";
-
 const rootReducer = combineReducers({
-  charactersReducer,
-  oneCharacterReducer,
   [characterApi.reducerPath]: characterApi.reducer,
 });
