+++ src/widgets/characterDetails/ui/CharacterDetails.tsx
@@ -1,21 +1,10 @@
 // import React from "react";
-import { useEffect } from "react";
-import {
-  useAppDispatch,
-  useAppSelector,
-} from "src/app/providers/store/config/hooks/redux";
-import { fetchCharacterById } from "src/app/providers/store/config/reducers/ActionCreators";
+
+import { useGetByIdQuery } from "src/shared/api/charactersApi";
 import { Loader } from "src/shared/ui/Loader/Loader";
 
 export const CharacterDetails = (prop: { id: number }) => {
-  const dispatch = useAppDispatch();
-  const { character, isLoading, error } = useAppSelector(
-    (state) => state.oneCharacterReducer,
-  );
-
-  useEffect(() => {
-    dispatch(fetchCharacterById(prop.id));
-  }, []);
+  const { data: character, error, isLoading } = useGetByIdQuery(prop.id);
 
   return (
     <div className="w-full my-8">
@@ -24,7 +13,7 @@ export const CharacterDetails = (prop: { id: number }) => {
           <Loader />
         </div>
       )}
-      {error && <h1>{error}</h1>}
+      {error && <h1>Произошла ошибка при загрузке</h1>}
 
       <div className="p-3 rounded-md bg-white shadow-md h-full;">
         <div className="grid grid-cols-2 gap-10">
