--- src/widgets/characterList/index.ts
@@ -1,1 +1,0 @@
-export { CharacterList } from "./ui/CharacterList";
